Overhaul Author Link schema
Author Links are now objects with a sort of subtype defined by their `site`
field. The node parser recognizes certain strings provided to the `links` field
and converts them to these objects pre-filled with the correct info (right now
it's Twitter, LinkedIn, and GitHub).

These objects can then be queried all throughout the site, whether it be for
social icons or a more detailed list.

Also worth noting that since `links` is an array, its order is maintained.