Paginate Author pages
I used `gatsby-awesome-pagination` here because it's easily the most well-known
Gatsby plugin for pagination, and using it confers more advantages over
home-brewing our own logic for it.

The blog still uses the original pagination, which I intend to change, but I
want to get this out ASAP so design work can progress.