Refactor header and fix banner+placeholder positioning issues
This commit primarily aims to make the Header placeholder grow and shrink while
accounting for the extra height of the alert banner, fixing many positioning
issues that were introduced with the banner addition.

This commit also refactors the logic of the Hamburger menu into a separate React
hook and a separate component for both the Hamburger menu and button.