Clean up broken links in content
The images in devops for data scientists are odd- they work on the site, but
don't actually point to real file path and break the link checker. I felt
changing them to a real path like other posts would be a better solution than
relying on an inconsistent usage of magic paths.

The GitHub link 404'ed because the paths were capitalized at "GitHub", while the
real path was a case sensitive "github".

Also a link to CML had a semicolon instead of a colon.