Fix details triggers that contain markup (#1736)
* Disable `simpleLinker` in headings and fix complex `details` triggers

Code quotes in headers are no longer auto-linked by `simpleLinker`, both for UX
and bugfix purposes.

Before, any markup like code quotes would break a header as only the first child
had its text extracted. Now the trigger uses the whole children value of the
header it's made from.

* Remove unused util import and unncessesary null assertion

* Change TS cast for triggerChildren to reflect that it's a ReactNode

* Remove dead code from GitHub Model

This is unrelated to the PR, but while we're ts-linting we may as well

* Simplify how details triggers/headings are handled

Now it just runs off the first (filtered) child. If that child is a header, it's
assumed the last item of the heading is an auto-link and removed.

Currently, every `details` first child is a heading, and all headings have
auto-links.

* Change details DVC-file to `.dvc` file

This is mostly as an example of this PR's fix working.

* Re-word comment and change heading type regex to use ^ and $

* Add JSX.Element cast to firstChild to quell TS

Even if the element is a string, the regex test will return false from a
non-string property.

* Refactor to only accept details with headings as the first child

* Fix malformed slice from refactor