Fix wording of copy-on-write notes for git-annex (#4127)
* Fix wording of copy-on-write notes for git-annex

The existing wording suggested that copy-on-write is safe because other copies get updated on write. In fact it's the opposite: unlike symlinks, where an edit would also change the original, copy-on-write is a safe means of deduplicating content because you can write to the copy without changing the original. The new wording clarifies this.

* Apply suggestions from code review

Co-authored-by: Jorge Orpinel <jorgeorpinel@users.noreply.github.com>